Added Scene Changer Script
Added a script to change scenes via their index number. This script has been applied to the next level button on the first puzzle. Germans in Russia level has been created however the image needs to be edited to include a desription and to make it fit properly on the screen.